chore(import): shared html-to-markdown + zip + foundation
Extract the Notion importer's hand-rolled jsdom block/inline renderer into a
generic apps/desktop/src/main/import/_shared/html-to-markdown.ts with
anchor/image/skipBlock hooks, and promote the zip-entry iterator to
_shared/zip.ts. Notion's convert-to-md now delegates to the shared converter
via hooks that reproduce its prior behaviour (page-link wikilinks, attachment
resolution, properties-table skip); Notion tests stay green. Other importers
(html, evernote, apple-journal, onenote, bear, google-keep) reuse these.
