+++ apps/desktop/src/main/import/notion/notion-importer.ts
@@ -3,7 +3,7 @@ import { createNote, updateNote } from '../../vault/notes-crud'
 import { saveAttachment } from '../../vault/attachments'
 import { createLogger } from '../../lib/logger'
 import type { Importer, ImportContext, ImportInput, ImportSummary } from '../types'
-import { forEachZipEntry } from './notion-zip'
+import { forEachZipEntry } from '../_shared/zip'
 import { parsePageInfo } from './parse-info'
 import { NotionResolverInfo } from './resolver'
 import { convertHtmlToMarkdown } from './convert-to-md'
